+++ Backend/serverconfig.js
@@ -1,0 +1,3 @@
+const PORT = process.env.PORT || 4000;
+
+module.exports = PORT;
