+++ src/Frontend/Components/Leaderboards/LeaderboardsComponents/DifficultyInput2.tsx
