+++ src/Frontend/Data/HTMLEntities.ts
@@ -1,0 +1,11 @@
+const entities = [
+    ["&quot;", '"'],
+    ["&#039;", "'"],
+    ["&rsquo;", "'"],
+    ["&amp;", "&"],
+    ["&ldquo;", '"'],
+    ["&eacute;", "é"],
+    ["&Uuml;", "Ü"],
+];
+
+export { entities };
